BUG: Newer versions of Numpy don't downcast floats to int
Failure in generating np.ones((ny, nx))

+++ pytopkapi/results_analysis/plot_soil_moisture_maps.py
@@ -106,8 +106,8 @@ def field_map_ndar(ndar_field,t,ar_coorx,ar_coory,X,image_out,variable):
     xmin=min(ar_coorx);xmax=max(ar_coorx)
     ymin=min(ar_coory);ymax=max(ar_coory)
     step=X
-    nx=(xmax-xmin)/step+1
-    ny=(ymax-ymin)/step+1
+    nx=int((xmax-xmin)/step+1)
+    ny=int((ymax-ymin)/step+1)
 
     ar_indx=np.array((ar_coorx-xmin)/step,int)
     ar_indy=np.array((ar_coory-ymin)/step,int)
